refact(videos): remodularize the views.
